backend/requirements.txt
absl-py==2.1.0
accessible-pygments==0.0.5
alabaster==1.0.0
altair==5.5.0
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.8.0
apispec==6.6.1
apispec-webframeworks==1.1.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
Automat==24.8.1
babel==2.16.0
beautifulsoup4==4.12.3
bidict==0.23.1
binaryornot==0.4.4
bleach==6.2.0
blinker==1.8.2
boto3==1.34.113
botocore==1.34.162
cachetools==5.5.2
certifi==2024.8.30
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorspacious==1.1.2
comm==0.2.2
constantly==23.10.4
contourpy==1.3.0
cookiecutter==2.6.0
cryptography==46.0.7
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
deepdiff==7.0.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docutils==0.21.2
et-xmlfile==1.1.0
executing==2.1.0
fastapi==0.115.11
fastjsonschema==2.21.1
filetype==1.2.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
fonttools==4.54.1
fqdn==1.5.1
frozendict==2.4.6
fsspec==2026.3.0
gast==0.6.0
gevent==24.2.1
gevent-websocket==0.10.1
gitdb==4.0.12
gitignore_parser==0.1.11
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-genai==1.73.1
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.74.0
greenlet==3.1.1
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.14.0
h2==4.3.0
h5py==3.11.0
hpack==4.1.0
httpcore==1.0.7
httplib2==0.31.2
httpx==0.28.1
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.9
imagesize==1.4.1
incremental==24.7.2
ipykernel==6.29.5
ipython==8.28.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.9.0
jmespath==1.0.1
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.5.0
kiwisolver==1.4.7
kthread==0.2.3
langchain-core==1.3.1
langchain-google-genai==4.2.2
langsmith==0.7.35
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.0
ml-dtypes==0.4.1
mmh3==5.2.1
mpl_sphinx_theme==3.9.0
multidict==6.7.1
multitasking==0.0.11
namex==0.0.8
narwhals==1.32.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
numpydoc==1.8.0
openai==1.66.3
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.12.1
ordered-set==4.1.0
orjson==3.11.8
overrides==7.7.0
packaging==26.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
peewee==3.17.9
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
postgrest==2.28.3
prometheus_client==0.21.1
prompt_toolkit==3.0.48
propcache==0.4.1
proto-plus==1.27.2
protobuf==5.29.6
psutil==6.1.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.13.3
pydantic_core==2.46.3
pydata-sphinx-theme==0.15.4
pydeck==0.9.1
Pygments==2.18.0
pyiceberg==0.11.1
PyJWT==2.12.1
pymongo==4.7.2
pyparsing==3.2.0
pyroaring==1.0.4
PyStemmer==2.2.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-json-logger==3.2.1
python-slugify==8.0.4
python-socketio==5.11.4
pytz==2024.1
pywin32==308
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
realtime==2.28.3
referencing==0.36.1
regex==2026.4.4
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.22.3
s3transfer==0.10.3
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==74.1.2
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
Sphinx==8.1.3
sphinx-copybutton==0.5.2
sphinx-gallery==0.18.0
sphinx-tags==0.4
sphinx_design==0.6.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sphinxcontrib-svg2pdfconverter==1.2.3
sphinxcontrib-video==0.2.1
SQLAlchemy==2.0.30
stack-data==0.6.3
starlette==0.46.1
storage3==2.28.3
streamlit==1.44.0
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
tabulate==0.9.0
taipy==4.0.0
taipy-common==4.0.0
taipy-core==4.0.0
taipy-gui==4.0.0
taipy-rest==4.0.0
taipy-templates==4.0.0
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-hub==0.12.0
tensorflow-intel==2.17.0
tensorflowjs==3.18.0
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
tiktoken==0.12.0
tinycss2==1.4.0
toml==0.10.2
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
Twisted==24.7.0
types-python-dateutil==2.9.0.20241003
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.2
tzlocal==5.2
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.2.3
uuid_utils==0.14.1
uvicorn==0.34.0
watchdog==4.0.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.4
wheel==0.44.0
widgetsnbextension==4.0.13
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.6.0
yarl==1.23.0
yfinance==0.2.55
zope.event==5.0
zope.interface==7.1.0
zstandard==0.25.0
